@@ -7,15 +7,17 @@
     "build": "vite build",
     "preview": "vite preview",
     "prepare": "npm run build",
-    "i18n": "i18next 'lib/**/*{.js,.jsx}'",
-    "lint": "standard"
+    "i18n": "i18next 'lib/**/*{.js,.jsx,.ts,.tsx}'",
+    "typecheck": "tsc -p tsconfig.json --noEmit",
+    "lint": "eslint . --ext .js,.jsx,.ts,.tsx"
   },
-  "main": "./dist/cobrowse-agent-ui.umd.js",
-  "module": "./dist/cobrowse-agent-ui.es.js",
+  "main": "./dist/cobrowse-agent-ui.js",
+  "module": "./dist/cobrowse-agent-ui.js",
+  "types": "./dist/cobrowse-agent-ui.d.ts",
   "exports": {
     ".": {
-      "import": "./dist/cobrowse-agent-ui.es.js",
-      "require": "./dist/cobrowse-agent-ui.umd.js"
+      "types": "./dist/cobrowse-agent-ui.d.ts",
+      "import": "./dist/cobrowse-agent-ui.js"
     },
     "./style.css": "./dist/cobrowse-agent-ui.css"
   },
@@ -38,14 +40,22 @@
     "not op_mini all"
   ],
   "devDependencies": {
+    "@microsoft/api-extractor": "^7.55.1",
+    "@types/react": "^19.2.7",
+    "@types/react-dom": "^19.2.3",
+    "@types/ua-parser-js": "^0.7.39",
     "@vitejs/plugin-react": "^5.1.1",
-    "eslint": "^8.23.1",
-    "eslint-plugin-react": "^7.31.8",
+    "cobrowse-agent-sdk": "^2.12.0",
+    "eslint": "^9.39.1",
+    "eslint-config-love": "^136.0.0",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^7.0.1",
     "i18next-parser": "^9.3.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "standard": "^17.0.0",
+    "typescript": "^5.9.3",
     "vite": "^7.2.4",
+    "vite-plugin-dts": "^4.5.4",
     "vite-plugin-svgr": "^4.5.0"
   }
 }
