@@ -1,14 +1,41 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react'
+import { useCallback, useEffect, useMemo, useState } from 'react'
 import {
   CodeEntry,
   Device,
+  type DeviceData,
   PlatformIcon,
   Session,
+  type SessionData,
   SmartConnectButton,
+  UserIcon,
   i18n
-} from '../lib/main.js'
+} from '../lib/main'
+import type { ReactNode } from 'react'
+import type { DeviceInfo } from 'cobrowse-agent-sdk'
 
-const deviceSamples = [
+interface SampleDeviceInfo {
+  platform: DeviceInfo['platform'],
+  device: string
+}
+interface SampleDevice {
+  id: string
+  name: string
+  location: string
+  online: boolean
+  connectable: boolean
+  last_active: number
+  device: SampleDeviceInfo
+}
+interface SampleSession {
+  id: string
+  state: SessionData['state']
+  recorded: boolean
+  activated: number
+  ended?: number
+  device: SampleDeviceInfo
+}
+
+const deviceSamples: SampleDevice[] = [
   {
     id: 'ios-1',
     name: 'Avery’s iPad Pro',
@@ -50,7 +77,7 @@ const deviceSamples = [
   }
 ]
 
-const sessionSamples = [
+const sessionSamples: SampleSession[] = [
   {
     id: 'session-1',
     state: 'active',
@@ -74,15 +101,44 @@ const sessionSamples = [
   }
 ]
 
-const platformList = ['web', 'ios', 'macos', 'android', 'windows', 'unknown']
+const platformList: DeviceInfo['platform'][] = ['web', 'ios', 'macos', 'android', 'windows']
+const users = [
+  { id: 'u1', name: 'Alex Johnson', colour: '#d3e4ff' },
+  { id: 'u2', name: 'Priya Patel', colour: '#ffe5b4' },
+  { id: 'u3', name: '李秀英', colour: '#f0d9ff' },
+  { id: 'u4', name: 'Chris', picture: 'https://avatars.githubusercontent.com/u/1?v=4' },
+  { id: 'u5', name: 'A B C D', colour: '#e0f7e9' }
+]
 
-const Section = ({ title, subtitle, children }) => (
+const Section = ({ title, subtitle, children }: { title: string, subtitle?: string, children: ReactNode }) => (
   <section className='demo-section'>
     <h2>{title}</h2>
     {subtitle ? <p>{subtitle}</p> : null}
     <div className='demo-grid'>{children}</div>
   </section>
 )
+
+const asDevice = (sample: typeof deviceSamples[number]): DeviceData => ({
+  online: sample.online,
+  connectable: sample.connectable,
+  last_active: new Date(sample.last_active),
+  device: {
+    platform: sample.device.platform,
+    device: sample.device.device
+  }
+})
+
+const asSession = (sample: typeof sessionSamples[number]): SessionData => ({
+  id: sample.id,
+  state: sample.state as SessionData['state'],
+  recorded: sample.recorded,
+  activated: new Date(sample.activated),
+  ended: sample.ended ? new Date(sample.ended) : null,
+  device: {
+    platform: sample.device.platform,
+    device: sample.device.device
+  }
+})
 
 export default function KitchenSink () {
   const [codeStatus, setCodeStatus] = useState('Enter 654321 to simulate a successful validation.')
@@ -92,23 +148,23 @@ export default function KitchenSink () {
   const [direction, setDirection] = useState('ltr')
 
   const languages = useMemo(() => {
-    const resources = (i18n.options && i18n.options.resources) || {}
+    const resources = i18n.options?.resources ?? {}
     return Object.keys(resources)
   }, [])
 
   useEffect(() => {
-    const handleLanguageChanged = (lng) => setLanguage(lng)
+    const handleLanguageChanged = (lng: string) => setLanguage(lng)
     i18n.on('languageChanged', handleLanguageChanged)
     return () => i18n.off('languageChanged', handleLanguageChanged)
   }, [])
 
   useEffect(() => {
-    const previousDir = document.documentElement.getAttribute('dir') || 'ltr'
+    const previousDir = document.documentElement.getAttribute('dir') ?? 'ltr'
     document.documentElement.setAttribute('dir', direction)
     return () => document.documentElement.setAttribute('dir', previousDir)
   }, [direction])
 
-  const handleCode = useCallback(async (code) => {
+  const handleCode = useCallback(async (code: string) => {
     setCodeStatus(`Checking ${code}...`)
     await new Promise((resolve) => setTimeout(resolve, 700))
     const success = code === '654321'
@@ -116,19 +172,19 @@ export default function KitchenSink () {
     return success
   }, [])
 
-  const handleConnect = useCallback((device) => {
+  const handleConnect = useCallback((device: DeviceData) => {
     const timestamp = new Date().toLocaleTimeString()
-    setLastConnection(`${timestamp} • Triggered connect(${device.name || device.device.platform})`)
+    setLastConnection(`${timestamp} • Triggered connect(${device.device.platform})`)
   }, [])
 
-  const handleRecording = useCallback((session) => {
+  const handleRecording = useCallback((session: SessionData) => {
     const timestamp = new Date().toLocaleTimeString()
     setRecordingInfo(`${timestamp} • Open recording for session ${session.id}`)
   }, [])
 
-  const handleLanguageChange = useCallback((event) => {
+  const handleLanguageChange = useCallback((event: React.ChangeEvent<HTMLSelectElement>) => {
     const next = event.target.value
-    i18n.changeLanguage(next)
+    void i18n.changeLanguage(next)
     setLanguage(next)
   }, [])
 
@@ -197,24 +253,28 @@ export default function KitchenSink () {
         subtitle='Devices highlight their online status and accept arbitrary children, such as SmartConnectButton.'
       >
         <div className='device-list'>
-          {deviceSamples.map((device) => (
-            <div key={device.id} className='device-card'>
-              <header>
-                <div>
-                  <h3 className='device-name'>{device.name}</h3>
-                  <div className='device-location'>{device.location}</div>
-                </div>
-                <span className='status-pill'>{device.online ? 'Online' : 'Offline'}</span>
-              </header>
-              <Device device={device}>
-                <SmartConnectButton
-                  device={device}
-                  label={device.connectable ? 'Connect' : 'Unavailable'}
-                  onClick={handleConnect}
-                />
-              </Device>
-            </div>
-          ))}
+          {deviceSamples.map((sample) => {
+            const device = asDevice(sample)
+            return (
+              <div key={sample.id} className='device-card'>
+                <header>
+                  <div>
+                    <h3 className='device-name'>{sample.name}</h3>
+                    <div className='device-location'>{sample.location}</div>
+                  </div>
+                  <span className='status-pill'>{sample.online ? 'Online' : 'Offline'}</span>
+                </header>
+                <Device device={device}>
+                  <SmartConnectButton
+                    device={device}
+                    onClick={handleConnect}
+                  >
+                    {device.connectable ? 'Connect' : 'Unavailable'}
+                  </SmartConnectButton>
+                </Device>
+              </div>
+            )
+          })}
         </div>
         {lastConnection ? <div className='log'>{lastConnection}</div> : null}
       </Section>
@@ -224,14 +284,18 @@ export default function KitchenSink () {
         subtitle='Demonstrates disabled vs. connectable styling outside of Device.'
       >
         <div className='button-row'>
-          {deviceSamples.map((device) => (
-            <SmartConnectButton
-              key={`standalone-${device.id}`}
-              device={device}
-              label={device.connectable ? `Connect ${device.device.platform}` : 'Not connectable'}
-              onClick={handleConnect}
-            />
-          ))}
+          {deviceSamples.map((sample) => {
+            const device = asDevice(sample)
+            return (
+              <SmartConnectButton
+                key={`standalone-${sample.id}`}
+                device={device}
+                onClick={handleConnect}
+              >
+                {device.connectable ? `Connect ${device.device.platform}` : 'Not connectable'}
+              </SmartConnectButton>
+            )
+          })}
         </div>
       </Section>
 
@@ -240,16 +304,39 @@ export default function KitchenSink () {
         subtitle='Shows active vs. recorded/ended sessions with Stopwatch durations.'
       >
         <div className='sessions'>
-          {sessionSamples.map((session) => (
-            <Session
-              key={session.id}
-              session={session}
-              className='panel'
-              openRecording={session.recorded ? handleRecording : undefined}
-            />
-          ))}
+          {sessionSamples.map((sample) => {
+            const session = asSession(sample)
+            return (
+              <Session
+                key={sample.id}
+                as={sample.recorded ? 'button' : 'div'}
+                session={session}
+                className='panel'
+                onClick={sample.recorded ? () => handleRecording(session) : undefined}
+              />
+            )
+          })}
         </div>
         {recordingInfo ? <div className='log'>{recordingInfo}</div> : null}
+      </Section>
+
+      <Section
+        title='UserIcon'
+        subtitle='Initials fallback, custom colours, and pictures.'
+      >
+        <div className='user-grid'>
+          {users.map((user) => (
+            <div key={user.id} className='user-card'>
+              <UserIcon user={user} />
+              <div className='user-label'>
+                <div className='user-name'>{user.name || 'Unnamed'}</div>
+                <div className='user-meta'>
+                  {user.picture ? 'Picture' : 'Initials'}
+                </div>
+              </div>
+            </div>
+          ))}
+        </div>
       </Section>
     </>
   )
