@@ -1,31 +1,44 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
+import dts from 'vite-plugin-dts'
 import svgr from 'vite-plugin-svgr'
 import { dirname, resolve } from 'path'
 import { fileURLToPath } from 'url'
+import pkg from './package.json'
 
 const __dirname = dirname(fileURLToPath(import.meta.url))
 
 export default defineConfig({
   plugins: [
     react(),
-    svgr()
+    svgr(),
+    dts({
+      rollupTypes: true
+    })
   ],
+  resolve: {
+    alias: {
+      '@': resolve(__dirname, 'lib')
+    }
+  },
+  css: {
+    modules: {
+      generateScopedName: '[name]__[local]__[hash:base64:5]'
+    }
+  },
   build: {
     lib: {
-      entry: resolve(__dirname, 'lib/main.js'),
+      entry: resolve(__dirname, 'lib/main.ts'),
       name: 'CobrowseAgentUI',
-      fileName: (format) => `cobrowse-agent-ui.${format}.js`
+      formats: ['es']
     },
     rollupOptions: {
-      external: ['react', 'react-dom', 'react/jsx-runtime'],
-      output: {
-        globals: {
-          react: 'React',
-          'react-dom': 'ReactDOM',
-          'react/jsx-runtime': 'jsxRuntime'
-        }
-      }
+      external: [
+        'react/jsx-runtime',
+        'date-fns/locale',
+        ...Object.keys(pkg.dependencies),
+        ...Object.keys(pkg.peerDependencies)
+      ]
     },
     sourcemap: true,
     target: 'es2015'
